backend/requirements.txt
fastapi[all]
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
psycopg[binary]
apscheduler
pydantic
aiohttp
aiosmtplib
python-jose[cryptography]
redis[hiredis]
antlr4-python3-runtime
pytest
httpx
passlib[bcrypt] 
python-multipart  